Possibly fixed Arsenal Module in MP, see comments
I think I finally figured out why it was botched in MP but worked perfectly in SP.

- In SP units appear where they are at the very start of the mission, in MP, however, units are first spawned somewhere at coordinates [10,10,0], and then teleported to where they should have been at the start.
- Due to that, setPosASL command that was launched right after mission start, didn't actually have any effect on units in multiplayer, so the unit wasn't even teleported into the Arsenal Room.
- After I band-aided that, turned out that Arsenal Module was picking players up right after they spawned at [10,10,0] and then returning them back to the same pos upon exit from the Arsenal.
- The Arsenal itself straight up wouldn't work and give me "BIS_fnc_SetIdentity: the unit can't be null" error when I was trying to launch it too soon after mission start, even though, to my understanding, the unit was never null.

So, ultimately, the solution was to put a delay that would give enough time for units to spawn and teleport to their actual coordinates before Arsenal Module picks them up.

There were also a lot of "game became unresponsive" situations due to loading screens that were supposed to make things better and load faster, but instead made the game freeze. I removed all forced loading screens, but the problem of why those were put in place didn't go anywhere - custom filled Arsenal was taking up to several minutes to boot for the first time, it was unacceptable.

The game is executing code during loading screens when the mission is being loaded and started (so, before time > 0). The code is executed right to a point where first scheduled code is met, then the script is paused until the game is started (time > 0). The various 'sleep o.oo1" that were scattered all over the place acted as stoppers that were intentionally delaying the execution of the script until the game is started. That was needed for "proper" work of forced loading screens in SP, but since those were removed, I removed the sleeps as well, and good riddance.