Fixes in Godmode Synchronizer Module
Fixed inconsistent godmode if given to multiple units at the start of the mission. Apparently, they all had the same diag_frameNo as key and so the hashmap value was overwritten multiple times resulting in a random unit coming out on top almost every restart. Replaced this with a simple counter variable that gets incremented every time a new value is inserted into the hashmap